requirements.txt
Django==4.0.3
django-rest-swagger==2.2.0
djangorestframework==3.11.0
ipdb==0.13.9
pytz==2022.6
uritemplate==4.1.1
urllib3==1.26.13